requirements.txt
pandas
numpy
lark-parser
reportlab
 